- updated readme (not done by a long shot) - changed main.py -> ignoramus.py - added copyright notice to one of the files. More to come.
