+++ ignoramus/templates/__init__.py
@@ -1,3 +1,5 @@
+# Copyright (C) 2018, Enrico Tuvera Jr
+
 import os
 import os.path
 
@@ -19,6 +21,23 @@ def generate_table():
     # NOTE: this _may_ come back to bite me if the files somehow get out-of-order
     return dict(zip(generate_lang_names(generate_files()), generate_files()))
 
+def generate_gitignore(lang, output):
+    print('generating {} for {}'.format(output, lang))
+    table = generate_table()
+    filename = table[lang]
+    with open(os.path.join(os.path.dirname(__file__), filename), 'r') as F:
+        header = [
+            '# Output generated by Ignoramus v0.0.1\n',
+            '# Written by Enrico Tuvera Jr\n'
+        ]
+        lines = F.readlines()
+        total = header + lines
+        target_path = os.path.join(os.getcwd(), output)
 
-if __name__ == '__main__':
-    generate_table()
+        print('writing file to: {}'.format(target_path))
+
+        with open(target_path, 'w') as G:
+            G.writelines(total)
+
+    print('complete!')
+
